% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeview.R
\name{hbrc.aesthetic}
\alias{hbrc.aesthetic}
\title{Modelling of landscape view attractiveness}
\usage{
hbrc.aesthetic(lcm, chm, dsm, observerh = 1.6, resc = NA, subs = 120, ltb)
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{chm}{Canopy height model including vegetation height above the ground.}

\item{dsm}{Digital surface model.}

\item{observerh}{Height of observer above the ground in same units as chm and dsm.}

\item{resc}{Rescaling factor for reducing resolution. Integer or NA if no aggregation required.}

\item{subs}{Number of regularly-spaced locations to calculate viewshed from. 120 takes approximately 10 minutes on a normal laptop, so be careful.}

\item{ltb}{Lookup table as matrix with two columns. First column must include all categories of land cover included in lcm. Second column indicates land covers that contribute to landscape attractiveness, indicated by a 1.}
}
\value{
A SpatRaster stack including three layers; lascore = landscape aesthetics score within the viewshed, vsarea = viewshed area, combinedscore = multiplicative landscape aesthetic score.
}
\description{
This function models relative landscape attractiveness using a viewshed.
}
