% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lai.R
\name{hbrc.lai}
\alias{hbrc.lai}
\title{Modelling of Leaf Area Index from Landsat images}
\usage{
hbrc.lai(lcm, lsat, ltb)
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{lsat}{Pre-processed Landsat 8 image with bands 3, 4, 5, and 7.}

\item{ltb}{Lookup table as matrix with two columns. First column must include all categories of land cover included in lcm. Second column indicates land covers that contribute to tree canopy cover, indicated by a 1.}
}
\value{
A rast object with one layer indicating leaf area index.
}
\description{
This function models Leaf Area Index from Landsat
}
