% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients2.R
\name{hbrc.nutrient}
\alias{hbrc.nutrient}
\title{Modelling of nutrient retention using InVEST}
\usage{
hbrc.nutrient(lcm, dem, runoff, resc = NA, subs = 120, ltb, bpt)
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{dem}{Digital elevation model.}

\item{runoff}{Runoff potential in mm as a SpatRaster. Can be created using the hbrc.runoff.cn function.}

\item{resc}{Rescaling factor for reducing resolution. Integer or NA if no aggregation required.}

\item{subs}{Number of regularly-spaced locations to calculate viewshed from.}

\item{ltb}{Lookup table with two columns. First column must include all categories of land cover included in lcm. Second is binary indicator for whether the land cover is water.}

\item{bpt}{Table of parameters for InVEST, with specific names (see example).}
}
\value{
A SpatRaster stack including one layer indicating the proportion of nutrient load exported.
}
\description{
This function models runoff retained by all land covers and additionally separates the contribution of ecosystems.
}
\examples{
# Example of creating the bpt variable from information in data(eslookup)
 bpt<-data.frame(
 LULC_desc = eslookup$name,
 lucode = eslookup$lcdb,
 load_n= eslookup$n.load,
 eff_n= eslookup$n.retention,
 crit_len_n= 25,
 proportion_subsurface_n= 0)
}
