% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.R
\name{hbrc.pm10}
\alias{hbrc.pm10}
\title{Modelling of particulate matter air pollution removed by tree cover}
\usage{
hbrc.pm10(lcm, lai, pm10 = 22, Vd = 0.0064, ltb)
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{lai}{Leaf Area Index map.}

\item{pm10}{PM 10 concentration in air, (μg/ m3)}

\item{Vd}{Deposition velocity (m/s). Should already account for resuspension rate.}

\item{ltb}{Lookup table as matrix with two columns. First column must include all categories of land cover included in lcm. Second column indicates land covers that contribute to tree canopy cover, indicated by a 1.}
}
\value{
A rast object with two layers. First layer is pm removal per m2. Second layer is actual pm removal per grid cell. Units are kg per day.
}
\description{
This function models removal of particulate matter air pollution by tree leaves over 24 hours
}
