% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff-cn.R
\name{hbrc.runoff.cn}
\alias{hbrc.runoff.cn}
\title{Modelling of runoff retention using a curve number approach}
\usage{
hbrc.runoff.cn(lcm, hsg, ltb, precip, antecedent = "average")
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{hsg}{Map of Hydrologic Soil Group as a SpatRaster with numeric values indicating 1 = HSG A, 2 = HSG B, 3 = HSG C, 4 = HSG D.}

\item{ltb}{Lookup table as matrix with five columns. First column must include all categories of land cover included in lcm. The remaining four columns must give curve numbers for HSG A, B, C, and D respectively.}

\item{precip}{Precipitation value given either as a single numeric value or SpatRaster.}

\item{antecedent}{Character value describing antecedent soil moisture conditions, either "dry", "average", or "wet".}
}
\value{
A SpatRaster with one layer indicating the proportion runoff retained.
}
\description{
This function models runoff retained by all land covers and additionally separates the contribution of ecosystems.
}
