% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff-e1.R
\name{hbrc.runoff.e1}
\alias{hbrc.runoff.e1}
\title{Modelling of runoff retention based on MBIE Building Code Clause E1 Surface Water}
\usage{
hbrc.runoff.e1(lcm, stm, slo, ltb)
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{stm}{Soil type map in three broad classes indicating 1 = high infiltration, 2 = mid infiltration, 3 = low infiltration.}

\item{slo}{Slope map in units of % slope.}

\item{ltb}{Lookup table as matrix with four columns. First column must include all categories of land cover included in lcm. The remaining three columns must give curve numbers for high, mid, and low infiltration soils respectively.}
}
\value{
A SpatRaster with one layer indicating the proportion runoff retained.
}
\description{
This function models runoff retained by all land covers and additionally separates the contribution of ecosystems.
}
