% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{hbrc.shade}
\alias{hbrc.shade}
\title{Modelling of shade provided by vegetation cover}
\usage{
hbrc.shade(lcm, chm, ccutoff = 0, resc = NA, dsm, sv)
}
\arguments{
\item{lcm}{Land cover map including vegetation categories.}

\item{chm}{Canopy height model including vegetation height above the ground.}

\item{ccutoff}{Cutoff value for minimum vegetation height to consider as providing shade.}

\item{resc}{Rescaling factor for reducing resolution. Integer or NA if no aggregation required.}

\item{dsm}{Digital surface model including full canopy.}

\item{sv}{Azimuth and zenith of the sun. Matrix with two columns. Can be created from a sequence of dates/ times using insol; see Example.}
}
\value{
A rast stack including four layers, all indicate proportion of time shaded. First is the surface shade with vegetation. Second is surface shade without vegetation. Third is shade with vegetation including under canopy. Fourth is contribution of vegetation in providing shade.
}
\description{
This function models shade provided by vegetation cover.
}
\examples{
# Example of creating the sv variable from a vector of dates and location
sv<- insol::sunpos(insol::sunvector(
 c(insol::JD(seq(ISOdate(2019,02,01,8),ISOdate(2019,02,01,18),by='hour'))),
 latitude = -43.52,
 longitude = 172.64,
 timezone=13))
}
