% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-carbon.R
\name{hbrc.tree.c}
\alias{hbrc.tree.c}
\title{Modelling of biomass caron stocks in trees}
\usage{
hbrc.tree.c(trees, treeheight, treedbh = NA, method, maxheight = Inf)
}
\arguments{
\item{trees}{Tree locations as a SpatVector.}

\item{treeheight}{Vector of tree height. Must be same length as trees.}

\item{treedbh}{Vector of tree diameter at breast height. Must be same length as trees. Optional - tree dbh will be estimated from height if no values are provided.}

\item{method}{Method of approximation of carbon stock - currently either "beets" for Beets et al 2012 approach or "sm" for Schwendenmann and Mitchell 2014.}

\item{maxheight}{Maximum height allowable, can be used to constrain very high value.}
}
\value{
A SpatVector including estimates of carbon stocks.
}
\description{
This function models biomass carbon stocks in trees based on allometric equations.
}
